+++ src/main/java/uk/co/reecedunn/intellij/plugin/xquery/lexer/XQueryLexer.kt
@@ -64,6 +64,7 @@ private val KEYWORDS = mapOf(
         "ancestor" to XQueryTokenType.K_ANCESTOR,
         "ancestor-or-self" to XQueryTokenType.K_ANCESTOR_OR_SELF,
         "and" to XQueryTokenType.K_AND,
+        "andAlso" to XQueryTokenType.K_ANDALSO, // Saxon 9.9
         "any" to XQueryTokenType.K_ANY, // Full Text 1.0
         "array" to XQueryTokenType.K_ARRAY, // XQuery 3.1
         "array-node" to XQueryTokenType.K_ARRAY_NODE, // MarkLogic 8.0
@@ -193,7 +194,7 @@ private val KEYWORDS = mapOf(
         "order" to XQueryTokenType.K_ORDER,
         "ordered" to XQueryTokenType.K_ORDERED,
         "ordering" to XQueryTokenType.K_ORDERING,
-        "orElse" to XQueryTokenType.K_ORELSE, // Saxon 9.8
+        "orElse" to XQueryTokenType.K_ORELSE, // Saxon 9.9
         "paragraph" to XQueryTokenType.K_PARAGRAPH, // Full Text 1.0
         "paragraphs" to XQueryTokenType.K_PARAGRAPHS, // Full Text 1.0
         "parent" to XQueryTokenType.K_PARENT,